+++ src/main/java/com/ict/esg/EsgProjectApplication.java
@@ -1,4 +1,4 @@
-package Ict.esgProject;
+package com.ict.esg;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
