+++ src/main/java/com/ict/esg/config/WebConfig.java
@@ -1,4 +1,4 @@
-package Ict.esgProject.config;
+package com.ict.esg.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
