+++ src/main/java/com/ict/esg/model/EvaluationResult.java
@@ -1,4 +1,4 @@
-package Ict.esgProject.model;
+package com.ict.esg.model;
 
 import lombok.Data;
