+++ src/main/java/com/ict/esg/repository/EvaluationResultMapper.java
@@ -1,5 +1,7 @@
-package Ict.esgProject.repository;
+package com.ict.esg.repository;
 
+import com.ict.esg.model.EvalCat;
+import com.ict.esg.model.EvaluationResult;
 import Ict.esgProject.model.*;
 import org.apache.ibatis.annotations.*;
