+++ src/main/java/com/ict/esg/service/SignUpService.java
@@ -1,15 +1,13 @@
-package Ict.esgProject.service;
+package com.ict.esg.service;
 
-import Ict.esgProject.model.EnterprisesInfo;
-import Ict.esgProject.repository.EnterprisesInfoMapper;
+import com.ict.esg.model.EnterprisesInfo;
+import com.ict.esg.repository.EnterprisesInfoMapper;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
 import java.util.Map;
 
 @Slf4j
