+++ src/test/java/com/ict/esg/EsgProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package Ict.esgProject;
+package com.ict.esg;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
